Unmerged paths:
        deleted by them: l10n/de/browser/chrome/browser/aboutAccounts.dtd
        both modified:   l10n/de/browser/chrome/browser/browser.dtd
        both modified:   l10n/de/browser/chrome/browser/browser.properties
        both modified:   l10n/de/browser/chrome/browser/preferences/preferences.properties
        both modified:   mozilla-release/browser/app/Makefile.in
        both modified:   mozilla-release/browser/app/macbuild/Contents/MacOS-files.in
        both modified:   mozilla-release/browser/app/moz.build
        both modified:   mozilla-release/browser/base/content/aboutDialog.xul
        both modified:   mozilla-release/browser/base/content/browser.xul
        both modified:   mozilla-release/browser/components/about/AboutRedirector.cpp
        both modified:   mozilla-release/browser/components/customizableui/CustomizableUI.jsm
        both modified:   mozilla-release/browser/components/customizableui/content/panelUI.inc.xul
        both modified:   mozilla-release/browser/components/newtab/aboutNewTabService.js
        both modified:   mozilla-release/browser/components/preferences/in-content/main.js
        both modified:   mozilla-release/browser/components/preferences/in-content/main.xul
        both modified:   mozilla-release/browser/components/preferences/in-content/preferences.xul
        both modified:   mozilla-release/browser/components/sessionstore/content/aboutSessionRestore.js
        both modified:   mozilla-release/browser/config/version_display.txt
        both modified:   mozilla-release/browser/extensions/moz.build
        both modified:   mozilla-release/browser/installer/package-manifest.in
        both modified:   mozilla-release/browser/installer/removed-files.in
        both modified:   mozilla-release/browser/locales/en-US/chrome/browser/preferences/privacy.dtd
        deleted by them: mozilla-release/browser/themes/linux/menuPanel-help.png
        deleted by them: mozilla-release/browser/themes/linux/menuPanel-help@2x.png
        both modified:   mozilla-release/browser/themes/osx/jar.mn
        deleted by them: mozilla-release/browser/themes/osx/menuPanel-customize.png
        deleted by them: mozilla-release/browser/themes/osx/menuPanel-customize@2x.png
        deleted by them: mozilla-release/browser/themes/osx/menuPanel-exit-yosemite.png
        deleted by them: mozilla-release/browser/themes/osx/menuPanel-exit-yosemite@2x.png
        deleted by them: mozilla-release/browser/themes/osx/menuPanel-exit.png
        deleted by them: mozilla-release/browser/themes/osx/menuPanel-exit@2x.png
        deleted by them: mozilla-release/browser/themes/osx/menuPanel-help-yosemite.png
        deleted by them: mozilla-release/browser/themes/osx/menuPanel-help-yosemite@2x.png
        deleted by them: mozilla-release/browser/themes/osx/menuPanel-help.png
        deleted by them: mozilla-release/browser/themes/osx/menuPanel-help@2x.png
        both modified:   mozilla-release/browser/themes/shared/customizableui/customizeMode.inc.css
        deleted by them: mozilla-release/browser/themes/windows/menuPanel-help.png
        deleted by them: mozilla-release/browser/themes/windows/menuPanel-help@2x.png
        both modified:   mozilla-release/build/upload.py
        deleted by them: mozilla-release/devtools/client/aboutdebugging/components/aboutdebugging.js
        both modified:   mozilla-release/devtools/client/aboutdebugging/components/moz.build
        both modified:   mozilla-release/devtools/client/jar.mn
        both modified:   mozilla-release/modules/libpref/init/all.js
        both modified:   mozilla-release/security/sandbox/mac/SandboxPolicies.h
        both modified:   mozilla-release/storage/mozStorageService.cpp
        both modified:   mozilla-release/testing/profiles/prefs_general.js
        both modified:   mozilla-release/toolkit/components/telemetry/Histograms.json
        both modified:   mozilla-release/toolkit/components/viewsource/content/viewSource-content.js
        both modified:   mozilla-release/toolkit/mozapps/extensions/content/extensions.js
        both modified:   mozilla-release/toolkit/mozapps/extensions/internal/XPIInstall.jsm
        both modified:   mozilla-release/toolkit/mozapps/extensions/internal/XPIProvider.jsm
        both modified:   mozilla-release/toolkit/mozapps/extensions/internal/XPIProviderUtils.js
        both modified:   mozilla-release/toolkit/mozapps/extensions/moz.build
        both modified:   mozilla-release/toolkit/mozapps/installer/packager.mk

+++ mozilla-release/devtools/client/shared/devices.js
@@ -4,7 +4,6 @@
 
 "use strict";
 
-const { Task } = require("devtools/shared/task");
 const { getJSON } = require("devtools/client/shared/getjson");
 const { LocalizationHelper } = require("devtools/shared/l10n");
 const L10N = new LocalizationHelper("devtools/client/locales/device.properties");
@@ -43,58 +42,81 @@ const LOCAL_DEVICES = "devtools.devices.local";
 let localDevices;
 let localDevicesLoaded = false;
 
-// Load local devices from storage.
-let loadLocalDevices = Task.async(function* () {
+/**
+ * Load local devices from storage.
+ */
+async function loadLocalDevices() {
   if (localDevicesLoaded) {
     return;
   }
-  let devicesJSON = yield asyncStorage.getItem(LOCAL_DEVICES);
+  let devicesJSON = await asyncStorage.getItem(LOCAL_DEVICES);
   if (!devicesJSON) {
     devicesJSON = "{}";
   }
   localDevices = JSON.parse(devicesJSON);
   localDevicesLoaded = true;
-});
+}
 
-// Add a device to the local catalog.
-let addDevice = Task.async(function* (device, type = "phones") {
-  yield loadLocalDevices();
+/**
+ * Add a device to the local catalog.
+ * Returns `true` if the device is added, `false` otherwise.
+ */
+async function addDevice(device, type = "phones") {
+  await loadLocalDevices();
   let list = localDevices[type];
   if (!list) {
     list = localDevices[type] = [];
   }
-  list.push(device);
-  yield asyncStorage.setItem(LOCAL_DEVICES, JSON.stringify(localDevices));
-});
-exports.addDevice = addDevice;
 
-// Remove a device from the local catalog.
-// returns `true` if the device is removed, `false` otherwise.
-let removeDevice = Task.async(function* (device, type = "phones") {
-  yield loadLocalDevices();
+  // Ensure the new device is has a unique name
+  let exists = list.some(entry => entry.name == device.name);
+  if (exists) {
+    return false;
+  }
+
+  list.push(Object.assign({}, device));
+  await asyncStorage.setItem(LOCAL_DEVICES, JSON.stringify(localDevices));
+
+  return true;
+}
+
+/**
+ * Remove a device from the local catalog.
+ * Returns `true` if the device is removed, `false` otherwise.
+ */
+async function removeDevice(device, type = "phones") {
+  await loadLocalDevices();
   let list = localDevices[type];
   if (!list) {
     return false;
   }
 
-  let index = list.findIndex(item => device);
-
-  if (index === -1) {
+  let index = list.findIndex(entry => entry.name == device.name);
+  if (index == -1) {
     return false;
   }
 
   list.splice(index, 1);
-  yield asyncStorage.setItem(LOCAL_DEVICES, JSON.stringify(localDevices));
+  await asyncStorage.setItem(LOCAL_DEVICES, JSON.stringify(localDevices));
 
   return true;
-});
-exports.removeDevice = removeDevice;
+}
 
-// Get the complete devices catalog.
-let getDevices = Task.async(function* () {
+/**
+ * Remove all local devices.  Useful to clear everything when testing.
+ */
+async function removeLocalDevices() {
+  await asyncStorage.removeItem(LOCAL_DEVICES);
+  localDevices = {};
+}
+
+/**
+ * Get the complete devices catalog.
+ */
+async function getDevices() {
   // Fetch common devices from Mozilla's CDN.
-  let devices = yield getJSON(DEVICES_URL);
-  yield loadLocalDevices();
+  let devices = await getJSON(DEVICES_URL);
+  await loadLocalDevices();
   for (let type in localDevices) {
     if (!devices[type]) {
       devices.TYPES.push(type);
@@ -103,11 +125,19 @@ let getDevices = Task.async(function* () {
     devices[type] = localDevices[type].concat(devices[type]);
   }
   return devices;
-});
-exports.getDevices = getDevices;
+}
 
-// Get the localized string for a device type.
+/**
+ * Get the localized string for a device type.
+ */
 function getDeviceString(deviceType) {
   return L10N.getStr("device." + deviceType);
 }
-exports.getDeviceString = getDeviceString;
+
+module.exports = {
+  addDevice,
+  removeDevice,
+  removeLocalDevices,
+  getDevices,
+  getDeviceString,
+};
